move every component to the correct folder Login, Singup and Cities
